Feature Selection Technique Comparision
This project analyzes the effect of feature selection techniques on model prediction performance. It uses two methods, statistical (Chi-squared Test of Independence) and dimensionality reduction (Principal Component Analysis), to compare three models to the original study from which the data was acquired. The three models that were used for comparison are Logistic Regression, Support Vector Machine, and Random Forest. In addition, two other models were trained and tested as a first step to another project. These two models include LASSO Logistic Regression and a Classification Tree.

The use of machine learning in medical practices is becoming more prevalent in areas where there are delayed testing results and immediate diagnoses. In this case, antibiotics are prescribed for UTIs that are not confirmed for up to 48 hours after testing. Machine learning can be used to reduce the over-prescription of antibiotics. This project aims to aid the imperative need to find the best methods for model training to produce reliable outcomes.